Image loaded

+++ include/libpng_wrapper.h
@@ -21,7 +21,7 @@ class PNGFileWrapper: public ImageFile
         ~PNGFileWrapper();
         uint32_t getHeight();
         uint32_t getWidth();
-        bool loadImage(std::function<uint32_t* (uint32_t)> callback);
+        bool loadImage(std::function<uint8_t* (uint32_t)> callback);
         uint8_t getComponentSize();
         uint8_t getComponentCnt();
 };
