+++ src/libpng_wrapper.cpp
@@ -67,7 +67,7 @@ uint8_t PNGFileWrapper::getComponentSize() {
     return _bit_depth;
 }
 
-bool PNGFileWrapper::loadImage(std::function<uint32_t* (uint32_t)> callback) {
+bool PNGFileWrapper::loadImage(std::function<uint8_t* (uint32_t)> callback) {
     for (uint32_t y = 0; y < _height; y++) {
         _row_pointers[y] = (png_byte*)callback(y);
 //        _row_pointers[y] = (png_byte*)callback(y, png_get_rowbytes(_png,_info));
